requirements.txt
elevenlabs>=1.54.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
